improve: PHPStan baseline 125 → 0 (-125 suppressions, perfect score)

+++ app/Jobs/AlertGenerationJob.php
@@ -287,7 +287,7 @@ class AlertGenerationJob implements ShouldQueue
     /**
      * Calculate hours until next maintenance based on machine type and hours.
      */
-    private function getHoursUntilMaintenance(Machine $machine): ?float
+    private function getHoursUntilMaintenance(Machine $machine): ?int
     {
         // Maintenance schedules based on machine type
         $maintenanceSchedule = [
